Add unit tests for all crawlers that exists
These unit tests are to check crawlers against the current behaviour.
Since the current crawlers may have bugs, the current behaviour
may be strange and out of expectations.

I have put some remarks in the README.md files for some of the strange
cases.

This commit is to prevent unexpected behaviour changes of the crawlers.
The bugs of the crawlers will be dealt with later.