Fix Libertytimes parse problem in Travis CI
The original parse function doesn't run correctly in Travis CI
platform. The HTML is parsed as UTF8 as it should. This problem is
strange because it does not happen in all other platform I run the
tests.

Add minor change to the Crawler_Libertytimes::parse method so it
runs correctly in both my development environment and Travis CI.